add a real 'cancel' button to 'start simplified setup?'
i tried to rename the 'cancel' button to 'login manually'
as suggested at https://github.com/deltachat/interface/issues/33

however, playing around with this change a bit in real setup situations,
this does not really help on making things easier.

- the user sees a dialog opening that lacks an option to 'just close'
  it at the first glance, one feels stuck

- he has to understand that 'login manually' is the thing he already
  started in the previous dialog

therefore, i think a simple 'cancel' might not be perfect,
but according to 'login manually' the better choice.

however, maybe we can reword the text, saying more explicit what cancel does.

but even then, not sure if the user really reads the things,
and i think the text is already a bit long.
also, whatever the user chooses, there is not really a 'bad choice' -
in both decisions, the user will either succeed logging in or come back to the
intial screen.

also, thinking of ios, there the 'cancel' button
is always much more highlighted as the other ones -
having buttons named 'cancel' feels natural, but having a bold
'login manually' is a bit weird as the 'simplified' choice is probably
easier in most cases.
sure, this is ios and we can do things different there,
however, it is another point :)