show foreground notification unconditionally
the foreground notification is an important element
that avoids the system from killing the app.

in the past, we added the foreground notification when the screen turns off
and removed it when the screen turns on again -
the reason was that the notification does not annoy users this way.

meanwhile, however, it seems that this approach is no longer as reliable
as in the past. might be that we are killed before the screen turns off -
or that some handlers are not executed by some systems.

also meanwhile, even the "allow to run in background" resp.
"allow battery consumption" seems to be unreliable on various systems,
there are several reports eg. on stackoverflow that apps are killed anyway -
but a foreground notification is not known to be killed yet -
also as the api contract is about to avoid exactly that.

finally, this notification is also a good methor for users
to notice if the app gets killed -
so that they can blame the system then.

therefore, the annoying of the permanent notification
is probably the smaller drawback compared to
received messages delayed.