use auto-linkify together with custom-linkify
unfortunately, auto-linking such as Linkify.addLinks(WEB_URLS...)
replaces custom-links set before by Linkify.addLinks(custom-pattern).

calling custom-linking after auto-linking may be too inaccurate
as in case of bot-commands such as /xkcd_123456
are treated partly as a phone number.

skipping auto-linking is hard as there are lots of things to do to get a
similar experience, see #1600.

finally, we solve the problem by replacing URLSpan by some other span after.
in fact, we do that anyway to get a different long-click behaviour.

the only important thing is that the span-replacement
must not be derived from URLSpan.