Don't cut the document icon in quotes
I also deleted an unnecessary LinearLayout, so you should look at the
diff with --ignore-all-space

Firstly, I lifted quote_attachment_container out of the FrameLayout with
the thumbnail so that quote_attachment_container can be wrap_content
(and enlarge the quote if necessary) while the thumbnail still always
matches the size of the quote.

Then I made the document view itself a little smaller so it does not
enlarge the quote all the times.