do not use cached value to check for notifications
the 'inboxes' value contains the last lines of notification,
however, if the app is restarted for whatever reasons,
but notifications are persisted,
they do not contain all lines.

for the notification display, this is no big deal,
however, this cached value is not suitable
to check if there is a notification or not for a chat.

even if there is a notification, 'inboxes' may be empty,
this 'speed up' check was introduced
for a workaround of notification removal that requires lots of calls.
meanwhile, however, we have DC_EVENT_MSGS_NOTICED,
so we also remove this workaround.