Better profile and group picture selection by using attachment selector (#1837)
fix #1836

AvatarSelector is mostly copied from AttachmentTypeSelector. I first started to make it a subclass, but there were so many functions that had to be changed that I decided against it (esp. also the constructor, so I would have had to create another constructor to allow AvatarSelector to bypass the "normal" constructor). Could still be done though, of course.

I think (hope joy) that the commits are clear enough that this can be reviewed commit-by-commit. In the first commit, commenting out code means "Moved somewhere else"

I found it easier to look at the first commit using gitk than GitHub as GitHub doesn't show it as copied from AttachmentTypeSelector but as a new file. (of course you can also just review normally, going throug it file-by-file :)

* Don't show remove button if there is no image yet

* Remove unnecessary animation that just made things look worse

* Delete now-unused ClearProfileAvatarActivity

* For groups, also use the attachment-chooser-dialog

* Remove confirmation dialog for removing avatar (The user has to confirm anyway, and this was not correctly translated)