Remove Deaddrop chat (Android UI) (#1975)
See deltachat/deltachat-core-rust#2514

Fixes #1963
Fixes #1890

Before deleting a chat the user is asked for confirmation. Before blocking not, because blocking is not "destructive" (i.e. can be completely reverted by unblocking the contact). I'm wondering whether we should show a snackbar with an Undo button though, but we can probably wait for user feedback on this.

* Start removing deaddrop chat

* Generally works

* Fix the buttons

(it's not easy to not show one of the buttons in
ConstraintLayout; if problems turn up, we should probably
just not use ConstraintLayout)

* fix more things

* restartLoader() instead of notifyDataSetChanged() because the latter didn't do anything

* Archived badge

* Rename "Contact request" to "Request"

* oops

* Remove setStockTranslation(8, context.getString(R.string.chat_contact_request));

* Remove commented-out code, we probably won't need it

* Don't show "Delete" button for 1:1 chats

* Get rid of short flickering when declining a request

* Set the app's night mode to match the preference

This added 6ms to startup time, but I think that's OK, as
we will also be able to use more `-night` views in the future

* Make "Accept" text a bit brighter and easier to read in dark mode

* Check if the chattype is Group instead of members.length == 1