Remove RequiresApi(19) annotation on ScribbleActivity (#2033)
These annotations don't actually have an effect, but they lead to an
error (not a warning) being shown in Android Studio. It still compiles
fine, though.

The line `case ScribbleActivity.SCRIBBLE_REQUEST_CODE:` in
ConversationActivity was therefore shown as an error all the time, which
was really annoying.

The alternative would be to suppress the error in ConversationActivity.