fix behavior of system-back-button for webxdc
for webxdc, the system-back-button should do the same
as the back-button in the title bar - just close the webxdc.

without this commit, it loads the previous page, which is often unexpected -
eg. many games use reload() for simplicity,
so you would have press one time 'back' for each round :)

even if there may be situations where this may be useful,
we do not want to encourage webxdc to rely on the system-back-button -
eg. desktop or ios currently just do not have this button.

technically, this pr changes the default -
so, for all classes derived from WebViewActivity, the system-back-button will
now behave as the title-bar-back-button.

for the local-help, this behavior is changed explictly -
i think, for HTML-mails, the 'just close' approach is also better.