use display name for verified-by row
using display name and email address adds quite some noise,
esp. if the email addresses are long and non-telling,
this does not really add sth. to "security".

depending one the language, this issue may be even worse.

the idea of showing the address was only to "protect" against ppl
that set their name to "You" or "Me" - however, this trick is maybe a joke
but no security issue:

- the verifier is always a directly verified, known contact
- we do not regard verified contacts as enemies
- the "Me" or "You" is visible and annoying at quite more visible placed
- you can always tap the row and will see
  that this is just a contact with a weird name
  (iOS, eg. shows a > to clarify that the cell is tappable,
  we could do sth. like that one android, other colors or whatever,
  however, i would not do that beforehand without more reasonings and usecases)

all in all, showing a long, non-telling row here,
seems over the top and not regarding the whole story.