don't say "Done" when we don't know we're done
adding a shortcut is sometimes a request only,
the user has to confirm and/or there is a subsequent dialog.

so, when `requestPinShortcut()` returns successfully,
saying "Done" as the result of "Add to Home Screen" is often wrong
from the view of the user.

however, some feedback is still needed for the cases there are no dialogs.

saying "One moment..." covers both cases sufficiently.