use term 'Bio' instead of 'Signature'
'Signature' is very much related to email, sth. below the text,
this is not the case for Delta Chat, esp. not when using chatmail

'Status' implies that others may be aware of changes,
which is not the case as well.

therefore we'll try 'Bio' - which is also used by by 'Telegram', 'Github', 'Mastodon';
(whatsapp and signal use 'About', but we were failing to get a good hint for that :)