consolidate wording for "Message Info"
before, "Info" and "Message Details" were used for the same things synonymously,

this is an issue mostly in the context menus,
where eg. iOS had "Info" but desktop did not -
and android had an "i"-icon ...

but also, tapping 'Info' and then getting a different title
(not only a longer version) is not super-nice.

with this change, everything is "Info", making also the "i"-icon fit.
this is also consistent with WhatsApp.
Signal has the same mess as we before, at least on iOS.