tune down copying email address
copying email addresses to the clipboard
is not proper way to get in contact in the light of chatmail -
even if the account is non-chatmail, the receiver may use chatmail.
or the copied contact.

it is better to point ppl to use 'Share' or 'Attach Contact' therfore
and not offer failing/corner-case things in the primary menu.

therefore, we do basically the same what desktop is doing all the years:
no special function, however, if you select text (desktop)
or long tap it (android now), you can copy it.