cleanup account switcher
do not show 'Add Profile' in bold.

this looks less cluttered
and makes the "bold" of the selected account more outstanding:

before, there were two types of "boldness",
the add profile button (using bold only)
and the selected profile (using bold and a background),
this results in a bit cluttered, unsteady look,
esp. when there are few accounts and more bold than normal is shown.

also, this makes "bold" more outstanding,
as this attribute is given to one line only.
there is no need for the "add profile" to be that outstanding
(in contrast to eg. the "add chat" dialog)

it is a minor, sure :)