center attachment icons
this PR centers the icons in the attachment selector,
the right-aligned ones looks esp. odd when no experiments are enabled.

the default now looks good,
and when all experiments are enabled, things are perfect as well.
but even if only one experiment is enabled, things are better imo
(but we anyway should optimize for default and not for experimental options)